updated to 0.1.7

+++ src/DOM/DOMConstants.java
@@ -1,5 +1,5 @@
 package DOM;
 
 public interface DOMConstants {
-	public static final int FITRADIUS=3;
+	public static final double FITRADIUS=3.0;
 }
